Pass byte offsets to swizzle_N, not float offsets.
This saves an instruction per swizzle-component in ARM64.

Bug: skia:13676
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/618818
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>

+++ src/core/SkRasterPipeline.h
@@ -283,7 +283,7 @@ struct SkRasterPipeline_CopySlotsCtx {
 
 struct SkRasterPipeline_SwizzleCtx {
     float *ptr;
-    uint8_t offsets[4];  // values must be multiplied by raster_pipeline_highp_stride by caller
+    uint16_t offsets[4];  // values must be byte offsets (4 * highp-stride * component-index)
 };
 
 class SkRasterPipeline {
